+++ src/app/components/index.ts
@@ -5,3 +5,4 @@ export * from './planets';
 export * from './species';
 export * from './starships';
 export * from './vehicles';
+export * from './page-not-found';