+++ src/app/components/page-not-found/page-not-found.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'sw-page-not-found',
+  template: require('./page-not-found.component.html'),
+  styles: [
+    require('./page-not-found.component.scss')
+  ]
+})
+export class PageNotFoundComponent {
+  constructor() { }
+}