+++ src/app/models/nav-item.model.ts
@@ -1,4 +1,4 @@
 export class NavItem {
   public label: string;
-  public routeName: string;
+  public path: string;
 }