+++ src/polyfills.ts
@@ -1,5 +1,5 @@
-import 'es6-shim';
-import 'reflect-metadata';
+import 'core-js/es6';
+import 'core-js/es7/reflect';
 require('zone.js/dist/zone');
 
 if (process.env.ENV === 'production') {