+++ src/vendor.ts
@@ -3,18 +3,11 @@ import '@angular/platform-browser';
 import '@angular/platform-browser-dynamic';
 import '@angular/core';
 import '@angular/common';
+import '@angular/forms';
 import '@angular/http';
-import '@angular/router-deprecated';
+import '@angular/router';
 
 // RxJS
 import 'rxjs';
 
-// Angular2 Material
-import '@angular2-material/core';
-import '@angular2-material/toolbar';
-import '@angular2-material/icon';
-import '@angular2-material/sidenav';
-import '@angular2-material/button';
-
 // other vendors here...
-import '../node_modules/normalize.css/normalize.css';